+++ src/gpgsa.c
@@ -304,7 +304,7 @@ size_t nmeaGPGSAGenerate(char *s, const size_t sz, const nmeaGPGSA *pack) {
   }
 
   /* checksum */
-  chars += nmeaAppendChecksum(s, sz, chars);
+  chars += nmeaAppendChecksum(s, sz, (size_t) chars);
 
   return (size_t) chars;
 }
